Exported billData function forreuse in bills page.
